+++ mayan/apps/cabinets/apps.py
@@ -4,6 +4,7 @@ from django.utils.translation import ugettext_lazy as _
 from mayan.apps.acls.classes import ModelPermission
 from mayan.apps.acls.permissions import permission_acl_edit, permission_acl_view
 from mayan.apps.common.apps import MayanAppConfig
+from mayan.apps.common.classes import ModelQueryFields
 from mayan.apps.common.menus import (
     menu_facet, menu_list_facet, menu_main, menu_multi_item, menu_object,
     menu_secondary
@@ -91,6 +92,9 @@ class CabinetsApp(MayanAppConfig):
                 permission_cabinet_remove_document
             )
         )
+
+        model_query_fields_document = ModelQueryFields(model=Document)
+        model_query_fields_document.add_prefetch_related_field(field_name='cabinets')
 
         def get_root_filter():
             return {