+++ mayan/apps/metadata/apps.py
@@ -8,7 +8,9 @@ from mayan.apps.acls.classes import ModelPermission
 from mayan.apps.acls.links import link_acl_list
 from mayan.apps.acls.permissions import permission_acl_edit, permission_acl_view
 from mayan.apps.common.apps import MayanAppConfig
-from mayan.apps.common.classes import ModelFieldRelated, ModelProperty
+from mayan.apps.common.classes import (
+    ModelFieldRelated, ModelProperty, ModelQueryFields
+)
 from mayan.apps.common.menus import (
     menu_facet, menu_list_facet, menu_multi_item, menu_object, menu_secondary,
     menu_setup
@@ -154,6 +156,9 @@ class MetadataApp(MayanAppConfig):
         ModelPermission.register_inheritance(
             model=DocumentMetadata, related='metadata_type',
         )
+
+        model_query_fields_document = ModelQueryFields(model=Document)
+        model_query_fields_document.add_prefetch_related_field(field_name='metadata')
 
         SourceColumn(
             source=Document, label=_('Metadata'),